+++ flexiapi/app/Http/Controllers/Api/Admin/VcardsStorageController.php
@@ -63,7 +63,7 @@ class VcardsStorageController extends Controller
         $vcard->vcard = preg_replace('/\r\n?/', "\n", $vcardo->serialize());
         $vcard->save();
 
-        return $vcard->vcard;
+        return $vcard;
     }
 
     public function update(Request $request, int $accountId, string $uuid)
@@ -89,7 +89,7 @@ class VcardsStorageController extends Controller
         $vcard->vcard = preg_replace('/\r\n?/', "\n", $vcardo->serialize());
         $vcard->save();
 
-        return $vcard->vcard;
+        return $vcard;
     }
 
     public function destroy(Request $request, int $accountId, string $uuid)